requirements.txt
faster-whisper
openai
pyyaml
loguru
chromadb
pdfplumber
python-pptx
